Add function to display comments
