--- src/modules/showUI.js
@@ -1,38 +1,0 @@
-/* const mainPage = document.querySelector('.main');
-// show card dynamically
-const shoWUI = (title, image) => {
-  const div1 = document.createElement('div');
-  div1.className = 'card';
-  const div2 = document.createElement('div');
-  div2.className = 'image';
-  const imgs = document.createElement('img');
-  imgs.setAttribute('src', image);
-  imgs.setAttribute('class', 'imgs');
-  const div3 = document.createElement('div');
-  div3.className = 'nameMovie';
-  const pE1 = document.createElement('p');
-  pE1.className = 'namedMovies';
-  pE1.textContent = title;
-  const div4 = document.createElement('div');
-  div4.className = 'likess';
-  const iE1 = document.createElement('i');
-  iE1.className = 'fa-regular fa-heart';
-  const span1 = document.createElement('span');
-  span1.className = 'numberLiked';
-  const pE2 = document.createElement('p');
-  pE2.className = 'likecount';
-  pE2.textContent = 'likes';
-  const div5 = document.createElement('div');
-  div5.className = 'btnsComRres';
-  const btn = document.createElement('button');
-  btn.className = 'btnComment';
-  btn.textContent = 'comment';
-  div5.appendChild(btn);
-  div4.append(iE1, span1, pE2);
-  div3.append(pE1, div4);
-  div2.appendChild(imgs);
-  div1.append(div2, div3, div5);
-  mainPage.appendChild(div1);
-};
-
-export default shoWUI; */
